+++ app/services/forest_liana/resource_creator.rb
@@ -20,6 +20,10 @@ module ForestLiana
       rescue ActiveRecord::StatementInvalid => exception
         # NOTICE: SQL request cannot be executed properly
         @errors = [{ detail: exception.cause.error }]
+      rescue ForestLiana::Errors::SerializeAttributeBadFormat => exception
+        @errors = [{ detail: exception.message }]
+      rescue => exception
+        @errors = [{ detail: exception.message }]
       end
     end