+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Added
 - Filters - Add the before x hours operator.
 
+### Changed
+- Errors - Improve error handling on Record creation and update.
+
 ### Fixed
 - Router - Catch resources routing errors with a 404 response instead of a crash.